Refactor stdlibs to load embedded BIRs
