+++ stdlib/io/bal/Ballerina.toml
@@ -1,0 +1,4 @@
+[package]
+org = "ballerina"
+name = "io"
+version = "0.0.1"
