+++ stdlib/io/externs/io.go
@@ -14,7 +14,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package io
+package externs
 
 import (
 	"fmt"
